# examples(allocation): ensure message outlives host call