… examples/allocation/tinygo/testdata/greet.wasm