Also, incorporate guidance from TinyGo to use unsafe.String rather than unsafe.SliceHeader.