@evacchi 2023-05-05
rebuilt with `make build.examples.tinygo`

```
❯ tinygo version
tinygo version 0.27.0 darwin/arm64 (using go version go1.20.4 and LLVM version 15.0.7)
```

results

```
# without this commit
$ go test -run='^$' -bench '^BenchmarkAllocation.*' ./internal/integration_test/vs/compiler ./internal/integration_test/vs/interpreter -count=6 > old.txt
# with this commit
$ go test -run='^$' -bench '^BenchmarkAllocation.*' ./internal/integration_test/vs/compiler ./internal/integration_test/vs/interpreter -count=6 > new.txt
$ benchstat old.txt new.txt 
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/vs/compiler
                          │   old.txt   │              new.txt               │
                          │   sec/op    │   sec/op     vs base               │
Allocation/Compile-10       3.664m ± 0%   3.647m ± 1%        ~ (p=0.065 n=6)
Allocation/Instantiate-10   149.5µ ± 4%   151.5µ ± 3%        ~ (p=0.937 n=6)
Allocation/Call-10          1.562µ ± 0%   1.403µ ± 0%  -10.15% (p=0.002 n=6)
geomean                     94.91µ        91.87µ        -3.21%

                          │   old.txt    │               new.txt                │
                          │     B/op     │     B/op      vs base                │
Allocation/Compile-10       2.404Mi ± 0%   2.400Mi ± 0%  -0.16% (p=0.002 n=6)
Allocation/Instantiate-10   318.3Ki ± 0%   318.3Ki ± 0%       ~ (p=0.937 n=6)
Allocation/Call-10            48.00 ± 0%     48.00 ± 0%       ~ (p=1.000 n=6) ¹
geomean                     33.24Ki        33.22Ki       -0.05%
¹ all samples are equal

                          │   old.txt   │               new.txt               │
                          │  allocs/op  │  allocs/op   vs base                │
Allocation/Compile-10       1.824k ± 0%   1.830k ± 0%  +0.33% (p=0.002 n=6)
Allocation/Instantiate-10    812.0 ± 0%    812.0 ± 0%       ~ (p=1.000 n=6) ¹
Allocation/Call-10           5.000 ± 0%    5.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                      194.9         195.1       +0.11%
¹ all samples are equal

pkg: github.com/tetratelabs/wazero/internal/integration_test/vs/interpreter
                          │   old.txt   │              new.txt              │
                          │   sec/op    │   sec/op     vs base              │
Allocation/Compile-10       1.266m ± 3%   1.264m ± 0%       ~ (p=0.310 n=6)
Allocation/Instantiate-10   99.10µ ± 6%   99.35µ ± 4%       ~ (p=0.699 n=6)
Allocation/Call-10          32.92µ ± 0%   29.89µ ± 1%  -9.21% (p=0.002 n=6)
geomean                     160.4µ        155.4µ       -3.14%

                          │   old.txt    │               new.txt               │
                          │     B/op     │     B/op      vs base               │
Allocation/Compile-10       1.448Mi ± 0%   1.445Mi ± 0%   -0.21% (p=0.002 n=6)
Allocation/Instantiate-10   219.7Ki ± 0%   219.7Ki ± 0%        ~ (p=0.463 n=6)
Allocation/Call-10          2.223Ki ± 0%   1.944Ki ± 0%  -12.54% (p=0.002 n=6)
geomean                     89.79Ki        85.81Ki        -4.44%

                          │   old.txt   │               new.txt                │
                          │  allocs/op  │  allocs/op   vs base                 │
Allocation/Compile-10       1.065k ± 0%   1.070k ± 0%   +0.47% (p=0.002 n=6)
Allocation/Instantiate-10    756.0 ± 0%    756.0 ± 0%        ~ (p=1.000 n=6) ¹
Allocation/Call-10           97.00 ± 0%    85.00 ± 1%  -12.37% (p=0.002 n=6)
geomean                      427.4         409.7        -4.16%
¹ all samples are equal